+++ packages/pytest-simcore/src/pytest_simcore/services_api_mocks_for_aiohttp_clients.py
@@ -172,9 +172,6 @@ async def director_v2_service_mock(
         r"^http://[a-z\-_]*director-v2:[0-9]+/v2/computations/.*:stop$"
     )
     delete_computation_pattern = get_computation_pattern
-    projects_networks_pattern = re.compile(
-        r"^http://[a-z\-_]*director-v2:[0-9]+/v2/dynamic_services/projects/.*/-/networks$"
-    )
 
     get_services_pattern = re.compile(
         r"^http://[a-z\-_]*director-v2:[0-9]+/v2/dynamic_services.*$"
@@ -202,7 +199,6 @@ async def director_v2_service_mock(
         repeat=True,
     )
     aioresponses_mocker.delete(delete_computation_pattern, status=204, repeat=True)
-    aioresponses_mocker.patch(projects_networks_pattern, status=204, repeat=True)
 
     return aioresponses_mocker
