+++ services/web/server/src/simcore_service_webserver/studies_dispatcher/_redirects_handlers.py
@@ -18,7 +18,7 @@ from servicelib.aiohttp.requests_validation import parse_request_query_parameter
 from servicelib.aiohttp.typing_extension import Handler
 from servicelib.logging_errors import create_troubleshotting_log_kwargs
 
-from ..dynamic_scheduler.api import update_projects_networks
+from ..dynamic_scheduler import api as dynamic_scheduler_api
 from ..products.api import get_product_name
 from ..utils import compose_support_error_msg
 from ..utils_aiohttp import create_redirect_to_page_response
@@ -252,7 +252,9 @@ async def get_redirection_to_viewer(request: web.Request):
             file_params.download_link,
             product_name=get_product_name(request),
         )
-        await update_projects_networks(request.app, project_id=project_id)
+        await dynamic_scheduler_api.update_projects_networks(
+            request.app, project_id=project_id
+        )
 
         response = _create_redirect_response_to_view_page(
             request.app,
@@ -281,7 +283,9 @@ async def get_redirection_to_viewer(request: web.Request):
             service_info=_create_service_info_from(valid_service),
             product_name=get_product_name(request),
         )
-        await update_projects_networks(request.app, project_id=project_id)
+        await dynamic_scheduler_api.update_projects_networks(
+            request.app, project_id=project_id
+        )
 
         response = _create_redirect_response_to_view_page(
             request.app,
@@ -317,7 +321,9 @@ async def get_redirection_to_viewer(request: web.Request):
             ).STUDIES_DEFAULT_FILE_THUMBNAIL,
             product_name=get_product_name(request),
         )
-        await update_projects_networks(request.app, project_id=project_id)
+        await dynamic_scheduler_api.update_projects_networks(
+            request.app, project_id=project_id
+        )
 
         response = _create_redirect_response_to_view_page(
             request.app,
