+++ services/web/server/tests/integration/01/test_garbage_collection.py
@@ -113,9 +113,6 @@ async def director_v2_service_mock(
         r"^http://[a-z\-_]*director-v2:[0-9]+/v2/computations/.*$"
     )
     delete_computation_pattern = get_computation_pattern
-    projects_networks_pattern = re.compile(
-        r"^http://[a-z\-_]*director-v2:[0-9]+/v2/dynamic_services/projects/.*/-/networks$"
-    )
 
     mocker.patch(
         "simcore_service_webserver.dynamic_scheduler.api.list_dynamic_services",
@@ -134,7 +131,6 @@ async def director_v2_service_mock(
             repeat=True,
         )
         mock.delete(delete_computation_pattern, status=204, repeat=True)
-        mock.patch(projects_networks_pattern, status=204, repeat=True)
         yield mock
