+++ services/web/server/tests/unit/with_dbs/03/version_control/test_version_control_core.py
@@ -37,6 +37,7 @@ async def test_workflow(
     user_project: ProjectDict,
     aiohttp_mocked_request: web.Request,
     request_update_project: Callable[[TestClient, UUID], Awaitable],
+    mock_dynamic_scheduler: None,
     director_v2_service_mock: None,
 ):
     vc_repo = VersionControlRepository.create_from_request(aiohttp_mocked_request)
