+++ services/web/server/tests/unit/with_dbs/03/version_control/test_version_control_handlers.py
@@ -50,6 +50,7 @@ async def test_workflow(
     client: TestClient,
     user_project: ProjectDict,
     request_update_project: Callable[[TestClient, UUID], Awaitable],
+    mock_dynamic_scheduler: None,
     director_v2_service_mock: None,
 ):
     # pylint: disable=too-many-statements
